Attempt to handle the API's level omission behavior
First and foremost, this reverses the change in the previous commit which
completely ignores a player if they're not on the overall hiscores. Upon
further investigation, it appears that levels can still show up in the
CSV API even if they're unranked (this isn't true for the hiscores page).

Now, if levels are omitted from the API, we will try to patch them in.
This should allow levels updates without being in the hiscores, while
still avoiding negative diffs.

This change also tweaks the "check" command to handle missing levels.