Support tracking players across multiple guilds
Update the state so that a particular guild ID maps to a list of players and a tracking channel.
This requires adding a few data structures inside the state object to do forward and reverse mappings.
A few things could be improved, like determining all guilds and determining the bot's admin.
There's logic in place to migrate the legacy data to all guilds the client is in.
Also, rename all usages of "player" to "rsn" to be consistent.