Do major refactor of how hi-scores are modeled
Add new utility method fetchHiScores which parses data from the API in ways that are useful to the bot
(e.g. keep separate mapping of known levels vs levels with assumed defaults). In the state class, only
store levels/bosses that are explicitly known (regardless of what their values are).