Improve logic for "priming" player data when tracked
Due to the weird way hiscore status is stored in the state in a set to indicate "off", new players
were not in the set and thus considered "on". This resulted in some annoying initial notifications
when tracking a player. Now, when a player is first tracked their status will always be written to
PG and state, and a notification will not be sent for that first primer update.

Also, mark a player as active when they're first globally tracked. This serves the purpose of
(1) ensuring every single player has some sort of activity timestamp even if they've never had a real
update, and (2) giving new players a chance to have quick refreshes immediately without having to
wait to get off the inactive queue.