Move all 404 players to the archive queue
If a player is primed and sees a 404, set their timestamp to 4 weeks ago.
If the player has been active enough to not be on the archive queue yet they
see a 404, also set their timestamp to 4 weeks ago. By doing this, we can bump
them to the archive queue while also maitaining a timestamp we can potentially
use later for purging them from being tracked altogether.

The only loose end is that players with no timestamp at all need to be given a dummy timestamp.