Don't assume total XP is always positive for players
Forgot that the value is -1 if the player is off the overall hiscores, yet still on there for other skills.
This change adds temp logic to purge all non-positive XP values from PG and the state, and adjusts update
logic to only process total XP that is positive. This changes also makes it so that updates e.g. level gains
are also treated as "activity" since some users will be getting updates yet have -1 overall total XP.