Retain player refresh timestamps in PG
This data was only in-memory before, as it was only used for monitoring e.g. /details.
However, with large queue durations the problem arose of players consistently being in
the back of the queue after frequent reboots due to arbitrary ordering of PG and node maps.
This commit fixes this issue by saving player refresh timestamps and putting less-recently
refreshed players at the front of the queue on reboot, guaranteeing a more consistent queue
duration between reboots. Also, this change officially rephrases "last update" to "last refresh".