Fix PlayerQueue#next
This method previously could return null values due to the fact that it assumed
that all lower-tier queues would have values to fall back on. If the only populated
queue was the first one, then it would blindly defer to a lower queue and find nothing.
Now, it checks to see if the remaining queues are populated before passing it off.