Handle rollbacks in PPU PG queries
When writing to the PPU table, use the lesser of the existing and excluded base values.
This way, rollbacks are cleanly handled so that no extra logic outside of PG is needed.
When a rollback occurs, no writes to the PPU table are needed because the subsequent
update that's detected will either overwrite or coalesce with the existing PPUs.