Support negative virtual level diffs
In the case of virtual level rollbacks, a rollback to 99 means the level
will become missing from the diff. The validation that asserts that the
"after" is a superset of "before" thus had to be removed and instead just
treat the missing "after" value as the baseline value.

Also, reduce the negative diff strike threshold from 20 to 10 because waiting
20 refreshes for most players is a long time.