Track unconfirmed player display names
Rather than showing a warning when the RSN param of the /track command doesn't
match the sanitized RSN, track the param as the player's "unconfirmed" display name.
Once the display name is finally confirmed, clear the unconfirmed display name.
This should be much less annoying for users who are confused by the sanitized names.