Parse unranked boss/clue/activity hiscores data
The hiscores seems to have changed such that rows will be reported even if unranked.
Previously, any unranked row would've been missing data as well (e.g. score is -1).
Now, it appears that valid data can still be returned in a row even if the rank is -1.
This will enable much more data to be picked up and reported by the bot.